basic structure and packages
