+++ babel.config.js
@@ -1,3 +1,25 @@
 module.exports = {
   presets: ['module:@react-native/babel-preset'],
+  plugins: [
+    'react-native-reanimated/plugin',
+    [
+      'module:react-native-dotenv',
+      {
+        moduleName: '@env',
+        path: '.env',
+        safe: false,
+        allowUndefined: true,
+        verbose: false,
+      },
+    ],
+    [
+      'module-resolver',
+      {
+        alias: {
+          '@screens': './src/screens',
+          '@env': ['node_modules/react-native-dotenv'],
+        },
+      },
+    ],
+  ],
 };
