+++ package.json
@@ -10,27 +10,43 @@
     "test": "jest"
   },
   "dependencies": {
+    "@babel/eslint-parser": "^7.24.1",
+    "@react-native-async-storage/async-storage": "^1.23.1",
+    "@react-navigation/bottom-tabs": "^6.5.20",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
+    "@supabase/supabase-js": "^2.42.4",
     "react": "18.2.0",
-    "react-native": "0.73.6"
+    "react-native": "0.73.2",
+    "react-native-dotenv": "^3.4.11",
+    "react-native-elements": "^3.4.3",
+    "react-native-reanimated": "^3.8.1",
+    "react-native-safe-area-context": "^4.9.0",
+    "react-native-screens": "^3.30.1",
+    "react-native-url-polyfill": "^2.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@babel/preset-env": "^7.20.0",
     "@babel/runtime": "^7.20.0",
+    "@react-native-community/eslint-config": "^3.2.0",
     "@react-native/babel-preset": "0.73.21",
     "@react-native/eslint-config": "0.73.2",
     "@react-native/metro-config": "0.73.5",
     "@react-native/typescript-config": "0.73.1",
     "@types/react": "^18.2.6",
     "@types/react-test-renderer": "^18.0.0",
+    "babel-eslint": "^10.1.0",
     "babel-jest": "^29.6.3",
-    "eslint": "^8.19.0",
+    "babel-plugin-module-resolver": "^5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "jest": "^29.6.3",
-    "prettier": "2.8.8",
+    "prettier": "^2.8.0",
     "react-test-renderer": "18.2.0",
     "typescript": "5.0.4"
   },
   "engines": {
-    "node": ">=18"
+    "node": ">=20"
   }
 }
