+++ src/screens/Dashboard/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Text, View} from 'react-native';
+
+const Dashboard = () => {
+  return (
+    <View>
+      <Text>DASH</Text>
+    </View>
+  );
+};
+
+export default Dashboard;
