+++ tsconfig.json
@@ -1,3 +1,11 @@
 {
-  "extends": "@react-native/typescript-config/tsconfig.json"
+  "extends": "@react-native/typescript-config/tsconfig.json",
+  "compilerOptions": {
+    "strict": true,
+    "baseUrl": ".",
+    "paths": {
+      "@screens": ["src/screens"],
+      "@env": ["node_modules/react-native-dotenv"]
+    }
+  }
 }
